Add teast commands and some button functionality
Added commands to create and delete channels for a test team.

Added Discord event handlers to handle subscribing actions to a DiscordClient event with constraints and calling those actions if those constraints are satisfied.

Added ComponentInteractionHandlers that are instanstiated when the persistent component they are registered to is interacted with.

Added ability to create team channels and initialise with messages.

Added "Change tile" and "Submit evidence" buttons for each team, with functionality.

Added "View evidence" button for each team, with no functionality.