Add buttons to create/join a team.
Added a slash command that posts a message with buttons to create and join a team.

Added support for a Modal DEH.

Changed the default of the NumberOfAttachments DEH constraint to null to make it not be required.

Changed the summary of some ComponentInteractionHandlers to better match the component they are handling.