Add CSVReader and TaskTemplatePopulator
CSVReader reads and parses the "Tasks.csv" and "Task restriction.csv" files, throw appropriate exceptions and saves results to database.

CSVReader is not currently set up to save TaskRestrictions to the database and it's ability to save Restrictions is commented out.

TaskTemplatePopulator reads all images in the Easy, Medium and Hard folders in the "Tile images" folder and converts them to a csv as a template.