Add support for database calls in the ChangeTileButtonHandler
The handler now pulls from the database when displaying options to change tiles.

It currently uses tasks as the tiles have not been set up to support multiple copies of tasks.

It does not save the choices to the database yet as, again, the tiles are not completely set up.