Refactor ComponentInteractionHandler.TrySendUserTeamStatusErrorMessage
Removed the need to notify admins since the database is treated as the source of truth.

Inverted the return value, so true means on error ocured and a message was sent.

Renamed the method to state what is it doing.

It no longer concludes the interaction and that was doing more than the method should.

Removed an unecassary else if that was never getting ran.