Add ability for admins to react to evidence for verification.
Admins can react to messages in the evidence channels to either verify or reject evidence.

Event handlers now register their own OnEvent method instead of Bot as that didn't make sense.

Added DiscordMessageId to Evidence in the DB to check if a message that's being reacted to is associated with some evidence.

Change the DiscordComponentEmojis to DiscordEmojis in the BingoBotCommon to make them more universal.