Start refactor.
Split the CSVReader into two classes: CSVReader and CSVOperator, to separate their responsibilities.

Added components for the CSVReader such as CSVLine and CSVValue to further split responsibilities and to make them easy to reuse.

Added tests for CSVLines, CSVValues and the CSVReader.

Started tests for CSVOperators.

Need to finish creating the required CSVOperators.