Add a semaphore and headers and fix vertical row positions.
Added a semaphore to LeaderboardDiscord to ensure that the leaderboard image is created and posted in the correct order should it be requested multiple times simultaneously.

Each row is now correctly positioned.

Added a header for each column.

Renamed row components to cells.

There is a bug in the headers where the column sizes don't get updated. The image imaging systems will be refactored so this will be dealt with later.