Update schema and tests, bugfix and change role creation and deletion.
Fixed a bug in RequestDeleteTeam where it was referencing DiscordEntities.DiscordTeam instead of RSBingot.DiscordTeam.

Role creation was moved to DiscordTeam so it can save the id to the DB.

RequestDeleteTeam now gets the role from DiscordTeam.

Added channel ids, role id and Tile.IsComplete to the schema.

Changed Verified to IsVerified in the schema to match the DBContext.

Added channel ids and role id to MockDBSetup.

Removed duplicated Add_Team method in MockDBSetup.