Fix bugs for creating and joining a team if a user is already on one.
The buttons will now check if a use is on a team before attemping to do anything else. If they are, then a common response is posted and the interaction is concluded.

Added a utility class to easily send and update messages and responses with interactions.

Fixed a bug where UserRepository.Exists was returned the inverted value.