Add validation to a team name, update some reponses.
A team name must only contain letters and/or number and cannot exceed 50 (as stated in the DB) characters.

When submitting a team name, the response message is now ephemeral.

Changed the response message for running a command that uses the Request system from "Processing response." to "Processing request.".

Change the MessageUtilities to use InteractionCreateEventArgs so all interactions can be easily responded to.