Fix team button ids and update messages when there are no tasks/tiles.
Fixed a bug where team buttons were being registered with the incorrect id.

Added a message telling the user when there are no tiles for their team when they attempt to interact with a submit image button.

Updated the message for ChangeTileButtonHandler for when there are no tasks in the database to make it clearer and ephemeral.