Implement the csv commands, allow RunRequest to post multiple messages and fixed bugs in GetCompiledMessages.
Moved compiling messages from CSVOperator to MessageUtilities and fixed the bugs with it. Need to refactor to be cleaner.

Fixed some bugs in CSVOperators and the CSVReader regarding warning and error messages.

Added "name" column to Restrictions.

Operator and reader exceptions now derive from RSBingoException.

Added ability to get the whitelisted domains from WhitelistChecker.

Added cdn.discordapp.com as a whitelisted domain for images.

Removed unnecessary commands in CommandController.

Implemented the CSV commands in CommandController.

Updated CommandController.RunRequest to allow multiple messages to be returned from a request and forwarded to the user.