Fix BoardImage bugs and add ability to auto create a team's tiles.
Added a test for the board image so the bot doesn't have to be ran every time. This is a test for visual inspection only.

Updated Paths to have public paths and a FromResources helper method.

Added a try catch to requests in the CommandController to avoid an exception if the interaction is no longer available, i.e. the channel was deleted.

Added the ability for a team to have it's tiles created automatically using the first x number of tasks, ordered by row id.

Fixed bugs in the BoardImage that was causing task images to be displayed in the incorrect position and with the incorrect size.

Converted most of the properties in BoardImagePreferences to consts.