Allow message reading, fix slash command registration and deleted overwritten evidence.
DiscordIntents has been changed to allow the bot to read messages and get attachments. Even though this previously worked.

The description of a command was too long so the bot would not start. This has been fixed.

If a user submits evidence when there is already some submitted by them for the same tile and there is a message for it in the pending evidence channel, the message will now be deleted as it is no longer needed.