Partially fix a bug in SubmitImageButtonHandler.
If an instance is active, then one of the tiles gets completed, attempting to submit evidence for one of those tiles will conclude the interaction and prompt the user to press the button again.

The intention was to refresh the select menu, but it is not working. Details are outlined in the class.