Add ability to specify the start and raise an event for it.
A timer will be set for the specified start DateTime. If it is in the past then it will be a 1s timer. When the timer elapses and event will be raised.

When the competition starts, the team board messages will be updated to reflect how they should look (enabling buttons etc.).