Begin refactor.
Create a Result class to easily pass around information about the outcome of a process.

Remove permissions validation from requests as it should be processed beforehand.

RequestsBase handles semaphores and catching exceptions to about wet code.

DiscordRequestServices can send a request to Discord in a try-catch to avoid having to do this in every place that wants to send a request.

In the future DiscordRequestServices will automatically retry sending a request if it was not a fatal error (i.e. there was a network error so it makes sense to try again).