Add DynamicGrid and fix a few bugs.
Update Grid to index arrays rather than using GetElement().

Added DynamicGrid to use Lists instead of arrays, like Grid. It also uses a generic type rather than string, something Grid should probably do as well.

Update Message and IMessage to use DynamicGrid for components as the 2d IEnumerable can be hard to follow.

Fixed a bug in IInteractionHandler to use Task instead of void as a return type for Conclue() to be useable with async.

Fixed a bug in RequestBase where the logger was of type RequestDeleteTeam.