Split AddUserToTeamReques and add RequestsValidationExtensions ..
Split AddUserToTeamRequest into: request, validation and handler.

Added RequestsValidationExtensions to avoid WET code in validators. Made this an extension instead of a base class as the request the validator uses matters for the methods. Thus a base class couldn't work as the given request would not work for all methods. The extensions rely on the request inheriting from interfaces that describe it's properties.

Added interfaces for request properties for use with RequestsValidationExtensions for FluidValidation.

Added ability to get DiscordRole for a team from RequestsUtilities.