Convert RequestsValidationExtensions to Validator.
Converted RequestsValidationExtensions to Validator as it was annoying having to prefix methods with "this." and pass in an IDataWorker to a lot of them.
Requests no longer need to inherit from an interface for each property, i.e. IRequestWithDiscordUser.
Each Validator doesn't need to create it's own IDataWorker as the base class now does this.
Validation method can be less verbose, i.e. ValidateDiscordUserNotOnATeam => ValidateUserNotOnATeam. Since they now take typed Func args instead of relying on the request's property interface.