Add SelectComponentFactory, SelectComponentHandler  and continue to refactor ViewEvidenceButtonHandler.
Add StrippedConstraints to ComponentInteractionDEH for public use as opposed to Constraints as CustomId only needs to be set by the system (likely a factory) that directly registers the component. Should put DEHs in their own assembly and make ComponentInteractionDEH.Constraints internal.

Made the callbacks in SelectComponent events to the SelectComponentHandler can subscribe to them. Since the handler is create on interaction, these cannot be registered when the component is created.

Added SelectComponentFactory to automatically register the SelectComponentHandler and apply ComponentInteractionDEH.Constraints.

Continue to refactor ViewEvidenceButtonHandler.

Added SelectComponentHandler to contain the selection handling and not bloat other classes.

Made RequestHandler take a generic Result as certain requests may require it, like creating a team's board would return an image.

Removed ability to add successes, warnings and errors with a string in RequestHandler to make it easier to test and the type of the class passed in can be directly compared.