Add (I)Interaction(Reason) to send messages to Discord and update Handler(Reason)s.
Added InteractionSuccess, InteractionWarning and InteractionError (and their interface counterparts) which add a Message to their MetaData to be sent to Discord via DiscordInteractionServices.

Added IWarning for RequestHandler to take instead of Warning to be consistent with successes and errors.

Made success, warning and error (normal and interaction) base classes abstract so they can not be used generically. This makes testing easier as the types can be compared with what is expected.
Renamed the non-interaction ones' prefix from Process to Handler to better state what they're for. Also make then inherit their FluentResults counterpart to get extra functionality.

DiscordInteractionServices checks the MetaData of handler Results for Messages to be posted to Discord.