Move SelectComponent functionality to SelectComponentUpdater and SelectComponentHandler .
Split SelectComponent into SelectComponent and SelectComponentUpdater to make SelectComponent a DTO and the updating functionality internal.

SelectComponentHandler now determines if a page or items was selected, updates the SelectComponent accordingly then calls the appropriate method for the derived class to handle.

Changed DiscordInteraction to ComponentInteractionCreateEventArgs on IInteractionRequest to give more details (required in SelectComponentHandler.

Changed RSBingoBot.RequestHandlers namespace to RSBingoBot.Component_interaction_handlers.