Add RequestContext, MetaData and ValidationExtensions.
Added RequestContext to wrap requests to provide MetaData to them. This MetaData is used internally in a request base class, then the underlying request is passed to the derived class for processing. Allows for passing information such as ComponentInteractionCreateEventArgs, DiscordComponent etc.
This also removes the need to have this additional data on the request interface themselves which means consumers don't have access to information they can't/don't need to use. E.g. SelectComponent was required in ISelectComponentRequest, this was not filled out until the request was passed into SelectFactory yet consumers had access to it and were able to set it despite it being overridden.

Added ValidationExtensions to aid configuring validators with MediatR.