Add Message.Tag, remove constraints from factories and finish updating ViewEvidenceButtonHandler and ViewEvidenceButtonValidator .
Added Message.Tag to be able to delete a group of Messages created through a hierarchy of InteractionHandlers.

Fixed Button factory to match it's base class updates.

Made InteractableComponentFactory no longer take interaction constraints as arguments since these should be handled in the validator for the request.

Removed InteractionArgs as there is no need to add a DTO here.

Made InteractionHandler.Conclue virtual so it can be used as a deconstructor. For example, to delete messages no longer needed.

Finish refactoring ViewEvidenceButtonHandler and ViewEvidenceButtonValidator to the new systems.

Renamed CloseButtonRequest to ConclueInteractionButtonRequest as that is more appropriate for what it should be doing. Originally it was going to clean up messages as well, but that is now passed to InteractionHandler.Conclue where it should be.