Add DiscordMessageServices and InitialiseTeamBoardChannelRequest and co.
Added Name to IComponent to identify it when logging.

Added DiscordMessageServices to be able to send messages to channels. Added Send to MessageExtensions in Lieu.

Added InitialiseTeamBoardChannelRequest and co. Which creates the initial message in the channel containing the team's board and buttons. The board is yet to be added.