Update DiscordChannelServices and add DiscordTeamFactory.
Updated the way requests and logging work in DiscordChannelServices and added ability to get a channel.

Added DiscordTeamFactory to dedicate the creation so it's not muddle with everything else DiscordTeam does.

Renamed DiscordTeam to DiscordTeamOld.

Added DiscordTeam as a DTO.

I didn't touch Program.cs so Idk what happened there.