Add SingletonButtons, update Bot and fix Button.
Added SingletonButtons to create all the singleton buttons to ensure they are registered with handlers. They will be registered even if they have not been sent to Discord but the bot has no way of knowing if they exist on Discord unless the id is stored in the database but that's unnecessary.

Button now inherits IInteractable and the id has been moved to ButtonInfo.

Fixed a bug in Button where the id was always a guid.

Updated Bot to create DiscordTeams from existing Teams via the DiscordTeamFactory. Also injected SingletonButtons into it so that the instance is created and thus the buttons.