Add SubmitDropButtonDTO and DiscordMessageServices.
Changed Message.Id to Message.DiscordMessage to give more information, such as the channel sent in.

Added DiscordMessageServices to begin implementing IDiscordMessageServices.

Added IDiscordMessageServices.RegisterCreationHandler, IMessageCreatedRequest and MessageCreatedHandler to be able to register requests to respond to messages being sent to Discord.

Added DeleteMessageRequest.

Updated SubmitDropButtonHandler to pass SubmitDropButtonDTO to it's child components so they can set and get values without having to know about other handler's existence, and so the handler processing remains in Process and the handler can be cleaned up once that method has finished.

Added SubmitDropMessageRequest to respond to the user posting evidence in a message.