Update SubmitDropSubmitButtonHandler and fix and implement messaging systems.
Message now sets Message.DiscordMessage if it is created with a DiscordMessage.

Implemented broken and missing methods in DiscordMessageServices.

Removed message methods from DiscordServices.

SendMessageHandler now sets the Message.DiscordMessage.

SubmitDropMessageHandler now updates the original interaction message to contain the evidence url.

Moved the image validation from SubmitDropMessageValidator to BingoValidator to be able to be reused and possible changed to use ImageSharp, easily.

Updated SubmitDropSubmitButtonHandler to delete existing messages if the new evidence is overriding.
Sends the evidence message to Discord.