Update SubmitDropButtonRequest and co.
EvidenceType is passed into SubmitDropButtonRequest as the handler should not determine this. For example, the drop requires the competition to have started and the board to be verified, the handler should not know this.

Removed completed tile from SubmitDropButtonHandler select since evidence cannot be submitted for them.

SubmitDropSubmitButtonHandler no longer tries to deleted evidence messages that have been accepted or rejected as they are required to stay for historical purposes. Plus completed tiles cannot be submitted for so the accepted evidence would never be deleted anyway.