Add ability to set exception messages for request handlers.
Add ability to set error messages when certain exceptions occur during processing a request handler.

Made the TResult have the new() constraint in RequestHandlerBase to avoid the downcasting concern when returning an error.

RequestHandlerBase now invokes PostProcess after Process.

Updated DiscordHandler to be prepared for this new system (no exceptions are yet to be added).

Updated DatabaseRequestHandler to use this new system and added the known db exceptions.

Updated CreateTeamRoleHandler to return DiscordRole as expected.