Refactor RenameTeam request and update team channel creation.
Made ChannelInfo to be used as a dto for creating and renaming channels.

Made DiscordTeamChannelsInfo to be used as a dto for creating and renaming team channels.

Deleted individual channel creation requests since the dto is sufficient.

Updated CreateMissingDiscordTeamEntitiesHandler to create each channel explicitly using the new dto.

Added RoleName to DiscordTeam to have a common place for the name regarding creating and renaming the role.