Fix some Message and Modal building issues and add PostTeamSignUpChannelMessage request.
Fixed an issue in Message where getting a builder with an empty row of components would throw.

Began to implement the + operator for Message. Only with content for now.

When adding content to a Message, added the option to choose how many lines to prefix it with. Also added the option to add an IEnumerable of content.

Fixed an issue when creating the builder for a Modal; it now correctly adds the required CustomId.

Commented out logging from DiscordInteractionMessagingServices with a TODO to move it to LoggingBehaviour, where it belongs.

Added TeamSignUpChannel to DataFactory.

Began to implement CommandController. Added the InitializeTeamSignUpChannel command and enabled and updated registering commands.

Added PostTeamSignUpChannelMessage request.

Updated CreateTeamButtonHandler to correctly send the Modal to Discord.