Fix bugs for Message, GetChannelHandler, board message and CreateMissingDiscordTeamEntitiesHandler.
MessageExtensions Send and Delete methods are now async.

Updated GetChannelHandler to return an error if the channel could not be retrieved.

Renamed the RSBingoBot/BingoCommands folder to RSBingoBot/Commands.

Added a dict to DiscordTeam to track active instances. This allows for ease of giving the user team choices.

Fixed an issue in CreateMissingDiscordTeamEntitiesHandler where the team's board message was attempting to be set before being sent to Discord, thus the required DiscordMessage didn't exist.

Added content to the board message created in CreateTeamBoardMessageHandler as sending the message would throw otherwise. This content may not be needed when the board image is sent.

Added the DeleteTeam command, though it is not fully implemented.