Add ValidationResponseBehaviour, fix some MediatR bugs and update logging.
Removed DiscordInteractionExtensions as it is now redundant with the methods having move to a service.

Removed HandlerBehaviour as the handlers are automatically called by MediatR so having this behaviour was duplicating the handlers.

Added ValidationResponseBehaviour to notify the user of validation failures.

Updated MessageSuccess to allow logging if the message was sent but unable to be retrieved for extra information (original interaction message for example with a server error when retrieving).

Fixed a bug in SendInteractionMessageHandler that was causing it to wait for the interaction to timeout before sending the message.

Added SendInteractionMessageWarning in case the original response was able to be sent but then not retrieved for logging info.

Fixed a bug in ValidationExtensions that was causing the ValidationBehaviour to be registered twice.

Updated ValidationExtensions to automatically registered interaction requests with the ValidationResponseBehaviour.

Fixed a bug in AllDiscordTeamsChoiceProvider that was using the team object rather than the team's name.

Reordered the some DI registering with Program and fixed some issues with DI.

Fixed an issue in NewTeamNameValidator that was causing the default message to be sent along with the errors, which was not desirered.