Add InteractionResponseBehaviour, interaction keep alive messages and InternalError.
Added SendKeepAlive to DiscordInteractionMessagingServices to ensure an interaction does not timeout.

Added InteractionResponseBehaviour and moved interaction responses from InteractionHandler to here since this is the better place for it.
Having this behaviour ensures that all IDiscordResponses returned by all requests created by the parent interaction request are sent to the user via recursion. Previously the handlers had to check the request results and pass them along which was tedious and error prone.

Updated ValidationResponseBehaviour to only check for ValidationErrors as this is all it is for.

Added BadRequestCheck to DiscordHandler as a common method to catch BadRequestExceptions and check their error codes.

Added (optional) automatic keep alive message sending to InteractionHandler.

Created InternalError and made RequestHandlerBase return it if there was an unexpected exception, rather than just a message. This is so it can be sent to the user.

SendModalHandler now returns InternalError if the interaction has already been responded to so the user it aware the bot errored,

Added AddInteractionResponseBehaviour to ValidationExtensions.