Update sending and tracking interaction messages and logging.
For InteractionMessage:
    removed FollowupMessageId as it is no longer used.
    added IsKeepAliveResponse for use in ValidationResponseBehaviour.

Added InteractionResponseTracker to track original responses to interactions. This allows the InteractionResponseBehaviour to check for an auto response to decide if it should send a default response if one has not been provided.

Updated LoggingBehaviour to make the request logging clearer and slightly more detailed.

Updated DeleteTeamHandler to use the RequestServices correctly.

Made IDatabaseServices inherit IRequestService as it should do.

Created requests for sending an original interaction response and a followup response. The requests to send an interaction message now uses these and so does sending a modal. These have been created to keep the error handling in one place.