Update request exceptions, CreateTeamRole and DeleteTeam and fix some bugs.
Fixed a bug in ValidationBehaviour that would not release the semaphores if the validation failed.

Fixed a bug in RequestHandlerBase that would always return an empty result upon catching an exception.

Updated RequestHandlerBase to return a ExceptionError for an unexpected error to log the exception message.

IDatabaseServices is now registered with DI.

Updated CreateTeamRole request to use the request services correctly and respond to the user correctly.

Update DeleteTeam request to better use the request system.