Fixed bugs with interaction responses and add command requests.
Added RequestMetaData to store information on the RequestTracker set by the framework so the user does not need to implement the pre-existing properties inherited from request interfaces. For example, having to implement Component for every request that inherits IComponentRequest, this is now hidden in the framework but can still be accessed from an extension on the request.

Added ICommandRequest.

Added slash commands to DiscordInteractionServices.

Made InteractionResponseBehaviour a base class for InteractionHandlerResponseBehaviour and ValidationResponseBehaviour.

Added extension methods for requests to be able to access its tracker information easily without having to grab RequestsTrackers and then scour the tracker for the desired one.

Added a try-catch for ValidationBehavior to ensure it does not throw. If the validator throws then a ValidationInternalError will be returned to be sent to the user.

Changed IInteractionRequest to no longer require a generic since the interaction args are now stored in the metadata. This has simplified its use in quite a few places such as InteractionHandler and ValidationExtensions.

RequestHandlerBase now sets the result on the request tracker instead of the RequestRunner. This is crucial because the behaviours to send responses to the user were being ran before the result was being set, thus they were not responding correctly (mostly had no responses to use so just sent the default response).

Removed an uneccassary semi-duplicate of RequestRunner.RunRequest.

Added DeleteTeamCommandRequest to Program validator registration as it was supposed to be.

Added DeleteTeamCommand request.