Moved properties to metadata for  all request.
Updated RequestMetaData to take null key values. If null, the name of the type of the value will be used as the key. This allows us to remove the keys from request, making it nicer to use.

Removed metadata key from ICommandRequest, IModalRequest , IInteractionRequest and IComponentInteractionRequest.

Removed Component from IComponentRequest since it is now in RequestMetaData.

Component is now registered via a instead of InteractableComponentFactory since the property has been removed from IComponentRequest.

Updated DiscordInteractionServices to take in a component when registering IComponentInteractionRequest to be added to the metadata.
Also registered the metadata will null key names so the value type names are used by default since the metadata keys have been removed from the request interfaces.

Updated various requests that implemented the properties inherited from a request interface to no longer have those properties since they no longer exist on the request.