Added RequestTrackerBehaviour and overhauled RequestsTracker.
Added a try-catch to InteractionHandlerResponseBehaviour to ensure that if anything past this point the the chain errors, the user is notified and the error can be logged.

Made RequestsTracker now use "pending" and "active" trackers to represent request trackers that have entered the pipeline and ones that have entered their corresponding handler. These are private, unlike the previous Trackers, as it allows for more control since these have specific uses and allow for them to be changed easily if required.

Added RequestTrackerBehaviour to centralise tracking usage and to be able to return errors to the logging and interaction response behaviours.

LoggingBehaviour no longer tells the tracker the request has completed, this has been moved to RequestTrackerBehaviour.

Removed InteractionHandlersTracker and associated files as it is now handled by RequestsTracker.

Moved setting the request handler result from RequestHandlerBase to RequestTrackerBehaviour as the result may not neccessarily be from the handler, plus it should be kept in the now dedicated class tracking usage.

Moved adding trackers from RequestRunner to  RequestTrackerBehaviour.