Make interactions use unique requests and bug fix some behaviours.
Registering an interactable request now requires a Func that return the request, rather than just taking the request. This is so the request can be unique for each interaction (even if the parameters are the same) which makes tracking so much easier.

InteractionResponseBehaviour will now only notify the user of a single instance of InternalError. They don't need to know how many occurrences since they don't even know what the errors are.

LoggingBehaviour now checks for a tracker's existence before using it to avoid and exception that then causes nothing to be logged.

Added error checking to RequestTrackerBehaviour to avoid it throwing.

Added RequestTrackerCleanupBehaviour to removed a tracker once it's request has been completed. This operation must be the very last thing (since logging requires the tracker) so it must be its own behaviour.

Renamed ValidationExtensions to MediatRExtensions since this is more accurate for what the class has become.