Convert DiscordTeamFactory to requests and fix issues mostly relating to creating and joining a team.
Converted DiscordTeamFactory into requests.

Removed ability to send and delete messages using MessageExtensions since it uses a request service and this must be initialised with the calling request. This should only be done via the handler base class.

Added a check in RequestTracker to see if the parent tracker exists before getting it to avoid it throwing during construction.

Changed the request services to transient rather than singletons with DI since they are supposed to hold a reference to the calling request, thus a new instance is needed each time. Otherwise they will always keep a reference to the first request that called them.

Updated DiscordTeamChannelsInfo to use methods rather than properties to get a channel's info. This is because when a parent is required that was not created at the time DiscordTeamChannelsInfo was created, the parent is null. Thus the info is now created on demand so the caller can ensure the parent exists.

AddUserToTeam request now takes in a dataworker to reduce db calls as the top level request can create the dw and save changes. It also now correctly uses a request service and its success is an IDiscordResponse so the user is aware they're added to the team.

JoinTeamButtonValidator  now requires the user to not be on a team since a user can only be on one team at a time.

JoinTeamSelectHandler now correctly uses request services and saves the dw.

Deleted JoinTeamSelectSuccess as it wasn't used.

Updated RenameTeamHandler to use the new DiscordTeamChannelsInfo  methods.

Added a perspective option when checking is a user is on a team in BingoValidator.

Added the create team requests to DiscordTeamServices and IDiscordTeamServices and reordered some methods.