Fixed issues with JoinTeamButtonHandler and MessageExtensions and update interaction responding.
Fixed an issue in MessageExtensions where it would always throw if a select component was added. It was only supposed to throw if a select component was in the same row as another component.

Sending a keep alive message to an interaction now asks if it should be ephemeral. The first response after will update the keep alive message thus using its ephemeral status and not that of the new response builder.

Removed the default response from InteractionHandlerResponseBehaviour since it was a nuisance. A response could have been sent some other way (such as a request service) and required some response to be added even if it was empty which just added clutter.

Fixed an issue in AddUserToTeamHandler where it was using the Team from the dw that first added Team to the DiscordTeam thus was causing saving issued. This should be completely resolved later on by changing Team to TeamId in DiscordTeam so a dw must pull their own instance to use it, ensuring it's not that of another.

Fixed an issue in JoinTeamButtonHandler where it was trying to send a message with no content, thus it was not sending.

Since All existing DiscordTeams are loaded on start up, JoinTeamButtonHandler select menu now uses those rather than pulling from the db and creating a DiscordTeam for each.
Also cleaned up the method considerably.

JoinTeamButtonHandler now sends the response directly rather than adding a response for the InteractionHandler to send, since those InteractionHandler responses are now redundant and need to be removed.

Fiexd an issue in JoinTeamButtonHandler where it was trying to use the user from JoinTeamButtonRequest rather than caching the user. This caused an error since the by  the time the select component was the be interacted with, the request tracker no longer existed and so the user was not able to be obtained.