Updated error handling with creating a team and fixed an issue in CreateNewTeamHandler team ids.
Component now takes a generic for the DiscordComponent type so it's properties can be accessed based on the component type rather than the base class.

RequestsTracker.ActiveCount now has the IBaseRequest  constraint for the generic since this is the key in the dictionary hence no other type should be passed in here.

In Validator, renamed RequestHandlerInstanceExists to ActiveRequestInstances to be clearer what it is validating.
Updated comments.
Fixed an issue where it was incorrectly checking the max count.

Updated UpdateDatabaseSuccess message to work with both plural and non-plural amounts of items.

Fixed an issue in CreateNewTeamHandler where the team's row id was always 0 (the db had to be saved to first). Removed the team from the db if the Discord entities errored and added an error if so, so the parent request is aware.

CreateTeamModalHandler now only updates the db if creating a team was successful.