Update validation semaphores and database error response.
Updated Validator semaphores to use a 'get' method rather than a virtual property. This allows the request and RequestSemaphores to be passed through to get the semaphores that require information to be accessed (like team specific ones).

RequestHandlerBase exception responses now takes an IError rather than a string, this allows the response to be sent to the user if IDiscordResponse is inherited.

DatabaseRequestHandler now uses UpdateDatabaseRequestError rather than a string as the exception response so that it can be sent to the user.

Updated the team registration button validators to use semaphores so a user cannot attempt to join and create a team at the same time (well, they still can but one request will wait for the other to ensure no concurrency issues).

Changed UpdateTeam to teamDatabase semaphores in RequestSemaphores so they are more specific and work for a specific team rather than blocking every team which was unessassary and could cause frustration.