Fix team creation bugs and update deleting a team.
In DiscordInteractionServices, renamed RegisterCommand to RunCommand to better suit what it does.

Added GetRoleError for when the role that's trying to be retrieved does not exist.

Removed AllDiscordTeamsChoiceProvider as it can only get the choices on app startup, but since teams can be added/removed at runtime, this it useless.

DeleteTeamCommandHandler now takes a team name rather than the DiscordTeam so its existence can be validated.

CreateMissingDiscordTeamEntitiesHandler now takes a dw so it's parent class can save it all in one go to improve performance. Also means that's potentially one less db connection to establish.

Updated CreateNewTeamHandler to add the newly created team to DiscordTeam.ExistingTeams.

Fixed a bug in CreateNewTeamHandler where it was not saving the team to the db.

Added apostrophes around the team name in CreateNewTeamSuccess to be consistent with other messages containing a team name.

Fixed a bug in DeleteTeamHandler where it was redefining its request services.

Added the removal of a team from DiscordTeam.ExistingTeams in DeleteTeamHandler.

Fixed a bug in DeleteTeamSuccess where the message contained the team's type instead of its name.

Added the team database semaphore to DeleteTeamValidator.

Removed a TeamExists method in DeleteTeamValidator as it will no longer be used.